+++ CHANGELOG.md
@@ -16,6 +16,14 @@ The following changes are pending, and will be applied on the next major release
 
 ## Unreleased
 
+### New feature
+
+- Passes the v2 of the JSON-LD context for Access Grants when issuing and using the
+  `/derive` endpoint. This allows the server to return all Access Credentials understood
+  by the client, both v1 and v2. Clients prior to this version will issue and retrieve
+  Access Credentials with a v1 context (and will not be able to use features introduced
+  in v2).
+
 ### Bugfix
 
 - Type declarations have been altered so that internal type declarations are no longer